replyFromRecipient functionality enabled again
In an "Experiment API" a hook into getIdentityForHeader was implemented.
Here we have access to the reply-to-message's header and get the result
of getIdentityForServer.
The information is extracted and sent asynchronously to the background
script. Here the information is correlated with newly appearing compose
windows and processed there.

The "Experiment API" does not allow to call synchronously into the
background script, so a direct replacement of the sender identity
in "getIdentityForServer" can not be done accross the two scripts.

A potential alternative solution could convey all settings to the
"Experiment API" script and handle the matching for "replyFromRecipient"
completely within the "Experiment API" script.