Add configurable header fields...
for finding the original recipient from a message that we reply-to.

Add a configurable list of SMTP headers to be consulted for matching.
An optional occurence number could be added.
E.g.
received#3
... to consult the 3rd "Receive: .." header line in the SMTP message header.
Without occurence number, all headers of that type are used.

The selected header contents are prepended to the "recipients" field from the Thunderbird
MessageHeader structure. So the additional header have a higher priority when
searching for identitites.