Show identity label in selection.
Thunderbirds allows to add a label to an identity. If you use the same
name and mail address for multiple identities (but e.g. different
different reply-to), this is a nice feature to differentiate them. This
change shows the label in the option selection if it exists. If no label
is set, nothing changes.