Allow alias-match of manually entered address after initial reply-hint match
When opening a reply or forward window for a message, a match from
the "old" recipients has the highest priority (if enabled).

This was also the case if a new "to" address was entered, which matched the
alias-detection-List.

Now the "old" recipients match is prioritized-down after first use. So when
later entering an alias-detection-match, this will be used.