Fix issue with popup dialog on wrong display
Thunderbird does not display webextension popup dialogs on correct screen.
(example setup: having linux xrandr display extension where 2 monitors are
stiched to a larger X-display, running Thunderbird on right screen opens
mail messages on correct screen, but webextension popup dialogs are shown on
left screen)

A partial fix was provided in
https://bugzilla.mozilla.org/show_bug.cgi?id=1812262
This fix provides at least the capability to control the position of a popup
window.
It was introduced in Thunderbird 112 (beta track).
The fix is expected to be part of the next release version 115
(planned for approx. July 2023), but is not yet part of the current
release version 102.10

Correct Identity does now place the popup relative to the current window.
A Thunderbird Version >=112 is required to see the effect.

Correct Identity was sucessfully tested with Thunderbird 113.0b4.