remove defaults from settings before writing to storage
This addresses issue #56 and avoids quota error:
    QuotaExceededError: storage.sync API call exceeded its quota limitations.

The "settings" entry in the storage contained settings for all configured
accounts (configured in Thunderbird) and all configured identities. The size
of the JSON encoded data is limited to 8 KiBytes:
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/storage/sync#storage_quotas_for_sync_data

By removing all default entries, the number of existing accounts and
identities is no longer an issue.

Other change: strict_max_version is increased to 145.*